Extract //cuttlefish/files:file_device_id from utils:files
Bug: b/537016318
