+++ base/cvd/cuttlefish/common/libs/utils/files.h
@@ -29,7 +29,6 @@
 
 namespace cuttlefish {
 bool FileExists(const std::string& path, bool follow_symlinks = true);
-Result<dev_t> FileDeviceId(const std::string& path);
 Result<bool> CanHardLink(const std::string& source,
                          const std::string& destination);
 inline Result<bool> CanRename(const std::string& source,
